--- src/configs/services.yaml
@@ -1,7 +1,0 @@
----
-kind: workerService
-version: udx.io/worker-v1/service
-services:
-  - name: "server"
-    autostart: "true"
-    command: "node index.js"
